components配下のディレクトリ構造見直し
