+++ front/src/components/organisms/common/ErrMsg.vue
