+++ front/src/components/organisms/common/FlashMsg.vue
