+++ front/src/components/organisms/common/Footer.vue
