+++ front/src/components/organisms/common/Loading.vue
