+++ front/src/components/organisms/top/AboutStay.vue
